fix(fine-tuning): rendre le suivi d'un entraînement exact et pilotable
Lot 1 de l'audit du suivi de fine-tuning : les sept constats à correctif
court, avec leurs tests de non-régression.

Progression (S3). « epoch=0.97 » était tronqué à 0, valeur falsy en JS :
barre, compteur d'époques et loss restaient masqués pendant toute la
première époque — le premier tiers d'un run par défaut — pendant que
l'étape courante affichait « epoch 0/3 ». L'époque devient fractionnaire
de bout en bout (Double, colonne migrée en DOUBLE PRECISION), l'affichage
arrondit au supérieur, et la progression n'est masquée que si elle est
réellement inconnue. La barre et l'ETA du centre d'activité avancent en
continu par la même occasion.

Courbe (S1 partiel, S2, S4). Les losses par étape étaient toutes écrasées
dans un point par époque : 2 à 3 points pour un run de 3 époques, moins
que les 2 exigés pour tracer. Elles sont désormais indexées sur l'époque
de la ligne elle-même, donc un point par étape journalisée. Le sondage
REST fusionne au lieu de remplacer — il effaçait l'eval_loss déposée par
le SSE, seule courbe qui signale un sur-apprentissage. L'axe X passe en
type="number" : son domaine était inerte sur l'axe catégoriel par défaut.

Étapes (S6). Un job COMPLETED laissait son jalon final éteint. La règle
sort du JSX vers une fonction pure testée.

Pilotage (S15, S16). Bouton d'arrêt avec confirmation sur la page de
fine-tuning — l'API et le centre d'activité le permettaient déjà, l'écran
dédié non. Le refus 409 passe en ProblemDetail et son motif s'affiche au
lieu de « Request failed with status code 409 ».

Télémétrie (S12). PYTHONUNBUFFERED des deux côtés : la phase antérieure au
premier pas, dont le téléchargement du modèle de base, n'atteignait le
flux qu'après coup. Et au plus une ligne de barre de progression par
seconde : les rafraîchissements tqdm, terminés par \r donc découpés comme
des lignes, retournaient le tampon de 500 du diffuseur en quelques
secondes.

Restent ouverts, documentés dans l'audit : la phase réelle d'un échec
n'est toujours pas conservée (S5), le flux SSE ne porte pas de jobId (S9),
et ni les logs ni la série de loss ne survivent à un rechargement (S8).

+++ backend/src/main/java/fr/spectra/persistence/FineTuningJobEntity.java
@@ -28,7 +28,8 @@ public class FineTuningJobEntity {
 
     private int datasetSize;
     private String currentStep;
-    private Integer currentEpoch;
+    /** Fractionnaire : voir {@link FineTuningJob#currentEpoch()} (colonne migrée en DOUBLE). */
+    private Double currentEpoch;
     private Integer totalEpochs;
     private Double loss;
     private Double evalLoss;
@@ -45,7 +46,7 @@ public class FineTuningJobEntity {
 
     public FineTuningJobEntity(String jobId, String status, String modelName, String baseModel,
                                String parameters, int datasetSize, String currentStep,
-                               Integer currentEpoch, Integer totalEpochs, Double loss,
+                               Double currentEpoch, Integer totalEpochs, Double loss,
                                Double evalLoss, String outputPath, String reportPath, String error,
                                Instant createdAt, Instant completedAt) {
         this.jobId = jobId;
